requirements.txt
asgiref==3.7.2
Django==4.2.2
pip==21.3.1
setuptools==68.0.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.40.0
django-dotenv
django-filter~=23.2
django-import-export
django-countries